auth_service/requirements.txt
# python=3.11.3

fastapi==0.103.1
pydantic-settings==2.0.3
uvicorn[standard]==0.23.2

# SQL ORM and migrations
sqlalchemy==2.0.20
alembic==1.12.0
aiosqlite==0.19.0

httpx==0.25.0

# JWT token
python-jose[cryptography]==3.3.0

# Password hashing
passlib[bcrypt]==1.7.4
